Split mailbox cell support modules
