+++ crates/shelterwood-mailbox/src/cell/waker_slot.rs
@@ -1,0 +1,83 @@
+use std::{sync::Arc, task::Waker};
+
+use crate::{MailboxRuntime, capability::dispose, panic::PanicAccumulator};
+
+/// The only storage surface for a caller-owned waker.
+///
+/// Its value is private even from the parent mailbox module, and every
+/// mutating operation requires an effects sink, so replacing or taking an
+/// `Option<Waker>` and accidentally dropping it beside a guard does not
+/// type-check.
+#[derive(Default)]
+pub(super) struct WakerSlot(Option<Waker>);
+
+pub(super) enum WakerAction {
+    Wake,
+    DropInline,
+    Dispose(Arc<dyn MailboxRuntime>),
+}
+
+enum WakerEffect {
+    Wake(Waker),
+    DropInline(Waker),
+    Dispose(Arc<dyn MailboxRuntime>, Waker),
+}
+
+#[derive(Default)]
+pub(super) struct WakerEffects(Vec<WakerEffect>);
+
+impl WakerSlot {
+    pub(super) fn will_wake(&self, waker: &Waker) -> bool {
+        self.0
+            .as_ref()
+            .is_some_and(|registered| registered.will_wake(waker))
+    }
+
+    pub(super) fn replace(&mut self, waker: Waker, effects: &mut WakerEffects) {
+        if let Some(displaced) = self.0.replace(waker) {
+            effects.push(displaced, WakerAction::DropInline);
+        }
+    }
+
+    pub(super) fn take(&mut self, action: WakerAction, effects: &mut WakerEffects) {
+        if let Some(waker) = self.0.take() {
+            effects.push(waker, action);
+        }
+    }
+
+    pub(super) fn is_empty(&self) -> bool {
+        self.0.is_none()
+    }
+}
+
+impl WakerEffects {
+    pub(super) fn is_empty(&self) -> bool {
+        self.0.is_empty()
+    }
+
+    fn push(&mut self, waker: Waker, action: WakerAction) {
+        self.0.push(match action {
+            WakerAction::Wake => WakerEffect::Wake(waker),
+            WakerAction::DropInline => WakerEffect::DropInline(waker),
+            WakerAction::Dispose(runtime) => WakerEffect::Dispose(runtime, waker),
+        });
+    }
+
+    pub(super) fn flush(&mut self, panics: &mut PanicAccumulator) {
+        for effect in self.0.drain(..) {
+            match effect {
+                WakerEffect::Wake(waker) => panics.run(|| waker.wake()),
+                WakerEffect::DropInline(waker) => panics.run(|| drop(waker)),
+                WakerEffect::Dispose(runtime, waker) => {
+                    panics.run(|| dispose(&runtime, waker));
+                }
+            }
+        }
+    }
+}
+
+impl Drop for WakerEffects {
+    fn drop(&mut self) {
+        self.flush(&mut PanicAccumulator::default());
+    }
+}
